fix: address SonarCloud new-code findings from #124/#125
- LiveFilesStorageInfo.fetch(): split the combined (int)(long) cast
  Sonar flagged as an unnecessary nested cast into two statements.
- LiveFileStorageInfoTest/LiveFileInfoTest: hoist size() out of the
  assertThatThrownBy lambdas that also called get() — Sonar wants at
  most one call that can throw per lambda so a real failure can be
  attributed unambiguously.
- LiveFileStorageInfoTest: replace `.length` + `isEqualTo` with
  AssertJ's hasSize().
- SstPartitionerFactoryTest: rename the unused try-with-resources
  `opts` binding to the unnamed `_`, matching the existing convention
  elsewhere in the test suite (e.g. PropertyTest's `Snapshot _`).

Quality gate was already OK; these are the 5 new-code findings that
were actually in my own changes (the other ~79 open findings in this
project are pre-existing, in files this work never touched).

+++ core/src/main/java/io/github/dfa1/rocksdbffm/LiveFilesStorageInfo.java
@@ -68,8 +68,8 @@ public final class LiveFilesStorageInfo extends NativeObject implements Iterable
 			MemorySegment err = RocksDB.errHolder(arena);
 			MemorySegment listPtr = (MemorySegment) MH_GET.invokeExact(dbPtr, optionsPtr, err);
 			RocksDB.checkError(err);
-			int count = (int) (long) MH_COUNT.invokeExact(listPtr);
-			return new LiveFilesStorageInfo(listPtr, count);
+			long count = (long) MH_COUNT.invokeExact(listPtr);
+			return new LiveFilesStorageInfo(listPtr, (int) count);
 		} catch (Throwable t) {
 			throw RocksDB.wrapInvokeFailure("getLiveFilesStorageInfo failed", t);
 		}
